Add helper method to create FileUploadContext. Reorganize FileSubmission related extension methods and tests.
